Created basic training module
